+++ Modules/SharePointDsc/Examples/Resources/SPSearchAuthoritativePage/1 - AuthoritativePage.ps1
@@ -1,6 +1,6 @@
 <#
 .EXAMPLE
-    This example shows how to create a Search Authoratative Page
+    This example shows how to create a Search Authoritative Page
 #>
 
     Configuration Example 
@@ -13,10 +13,10 @@
         Import-DscResource -ModuleName SharePointDsc
 
         node localhost {
-            SPSearchAuthoratativePage AuthoratativePage
+            SPSearchAuthoritativePage AuthoratativePage
             {
                 ServiceAppName       = "Search Service Application"
-                Path                 = "http://site.sharepoint.com/Pages/authoratative.aspx"
+                Path                 = "http://site.sharepoint.com/Pages/authoritative.aspx"
                 Action               = "Authoratative"
                 Level                = 0.0
                 Ensure               = "Present"