+++ Tests/Unit/SharePointDsc/SharePointDsc.SPSearchAuthoritativePage.Mocks.cs
@@ -1,0 +1,9 @@
+namespace Microsoft.Office.Server.Search.Administration
+{ 
+    public enum SearchObjectLevel {
+        SPWeb,
+        SPSite
+        SPSiteSubscription,
+        Ssa
+    }
+}